# OpenVMS: cluster the bitmap and stream the volume so it grows past 2 GB
BITMAP.SYS is pinned at 16 blocks by the layout, which capped the volume at
the 65 536 LBNs one bitmap can address -- 32 MB. ODS-2 already answers this:
HM2$W_CLUSTER sizes the volume in clusters, so one bit covers a whole cluster
and the bitmap stays put however large the volume gets. The writer now picks
the smallest power-of-two cluster that fits, aligns every run to a cluster
boundary, and records it in the home block.

Runs must stay cluster-aligned on both ends of the search: the allocator's
cursor steps by whole clusters, because a run that started mid-cluster would
share its bit with the run before it and the allocator would hand the same
blocks out twice.

That lifted the LBN ceiling; the byte ceiling needed the rest:

- The volume is emitted as a metadata prefix (boot block, home block,
  BITMAP.SYS, INDEXF.SYS, root directory -- 275 blocks) followed by payloads
  copied into place, so Create never holds the image. Build(byte[]) keeps
  working below the array limit.
- LbnToByteOffset returns long. It was int * int, which wraps negative on any
  volume past 4 GB.
- Reading goes through OpenVmsVolume: the prefix carries everything needed to
  recognise and list a volume, and file contents come straight off the backing
  stream. The old path read the image through a 16 MB cap, so a 40 MB volume
  returned the right bytes for its first file and garbage for the rest.
- OpenEntry hands back a window onto the volume rather than a copy -- every
  file the writer and the in-place modifier emit is one contiguous run.

FULL.disk, metadata.ini and home_block.bin now surface only on foreign images.
On our own volumes they made every rebuild fold the header surface back in as
real files, which is why defrag refused the round-trip ("entry set 7 -> 10").

Verified with a 2.6 GB four-file payload: create, list+extract, shrink and
defrag all return the source bytes.

+++ Compression.Tests/OpenVms/OpenVmsInPlaceModifyTests.cs
@@ -44,8 +44,8 @@ public class OpenVmsInPlaceModifyTests {
     var betaLbn = preBetaFh.Extents[0].StartLbn;
     var preAlphaBytes = image.AsSpan(alphaLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray();
     var preBetaBytes = image.AsSpan(betaLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray();
-    var preAlphaFhBlock = image.AsSpan(OpenVmsLayout.FileHeaderByteOffset(preAlphaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
-    var preBetaFhBlock = image.AsSpan(OpenVmsLayout.FileHeaderByteOffset(preBetaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
+    var preAlphaFhBlock = image.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(preAlphaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
+    var preBetaFhBlock = image.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(preBetaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
 
     // Add a third file in place.
     using var ms = new MemoryStream();
@@ -63,9 +63,9 @@ public class OpenVmsInPlaceModifyTests {
       Is.EqualTo(preBetaBytes), "BETA's data LBN must be byte-identical after Add.");
 
     // Pre-existing File Headers untouched.
-    Assert.That(modified.AsSpan(OpenVmsLayout.FileHeaderByteOffset(preAlphaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
+    Assert.That(modified.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(preAlphaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preAlphaFhBlock), "ALPHA's FH bytes must be byte-identical after Add.");
-    Assert.That(modified.AsSpan(OpenVmsLayout.FileHeaderByteOffset(preBetaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
+    Assert.That(modified.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(preBetaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preBetaFhBlock), "BETA's FH bytes must be byte-identical after Add.");
 
     // And the new file round-trips.
@@ -96,8 +96,8 @@ public class OpenVmsInPlaceModifyTests {
 
     var preAlphaBytes = image.AsSpan(alphaFh.Extents[0].StartLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray();
     var preGammaBytes = image.AsSpan(gammaFh.Extents[0].StartLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray();
-    var preAlphaFhBlock = image.AsSpan(OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
-    var preGammaFhBlock = image.AsSpan(OpenVmsLayout.FileHeaderByteOffset(gammaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
+    var preAlphaFhBlock = image.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
+    var preGammaFhBlock = image.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(gammaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
 
     using var ms = new MemoryStream();
     ms.Write(image, 0, image.Length);
@@ -110,9 +110,9 @@ public class OpenVmsInPlaceModifyTests {
       Is.EqualTo(preAlphaBytes), "ALPHA's data LBN must be byte-identical after Remove.");
     Assert.That(modified.AsSpan(gammaFh.Extents[0].StartLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preGammaBytes), "GAMMA's data LBN must be byte-identical after Remove.");
-    Assert.That(modified.AsSpan(OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
+    Assert.That(modified.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preAlphaFhBlock), "ALPHA's FH bytes must be byte-identical after Remove.");
-    Assert.That(modified.AsSpan(OpenVmsLayout.FileHeaderByteOffset(gammaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
+    Assert.That(modified.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(gammaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preGammaFhBlock), "GAMMA's FH bytes must be byte-identical after Remove.");
 
     // The removed file's data LBN is zero-wiped.
@@ -120,7 +120,7 @@ public class OpenVmsInPlaceModifyTests {
     Assert.That(betaSlice, Is.All.EqualTo((byte)0), "Removed file's data LBN must be zero-wiped.");
 
     // The removed FH is marked free (struc-level word zero).
-    var betaFhBlock = modified.AsSpan(OpenVmsLayout.FileHeaderByteOffset(betaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
+    var betaFhBlock = modified.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(betaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
     var strucLev = (ushort)(betaFhBlock[OpenVmsLayout.FhStrucLev] | (betaFhBlock[OpenVmsLayout.FhStrucLev + 1] << 8));
     Assert.That(strucLev, Is.EqualTo(0), "Removed file's FH must have struc-level zero.");
 
@@ -143,7 +143,7 @@ public class OpenVmsInPlaceModifyTests {
     var pre = new OpenVmsReader(image);
     var alphaFh = OpenVmsReader.ReadFileHeader(image, pre.Entries[0].FileId)!;
     var preAlphaBytes = image.AsSpan(alphaFh.Extents[0].StartLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray();
-    var preAlphaFhBlock = image.AsSpan(OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
+    var preAlphaFhBlock = image.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray();
 
     using var ms = new MemoryStream();
     ms.Write(image, 0, image.Length);
@@ -155,7 +155,7 @@ public class OpenVmsInPlaceModifyTests {
     // ALPHA untouched.
     Assert.That(modified.AsSpan(alphaFh.Extents[0].StartLbn * OpenVmsLayout.BlockSize, OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preAlphaBytes), "ALPHA's data LBN must survive a BETA replace.");
-    Assert.That(modified.AsSpan(OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
+    Assert.That(modified.AsSpan((int)OpenVmsLayout.FileHeaderByteOffset(alphaFh.FileId), OpenVmsLayout.BlockSize).ToArray(),
       Is.EqualTo(preAlphaFhBlock), "ALPHA's FH bytes must survive a BETA replace.");
 
     // New BETA round-trips with the new bytes.
